+++ packages/integrations/components/integrationsManager/componentMap.js
@@ -1,0 +1,7 @@
+// Integration components.
+import GoogleAnalytics from '../googleAnalytics';
+
+// Declare a map of available integration components with a corresponding key.
+export default [
+  { key: 'googleAnalytics', el: GoogleAnalytics },
+];