+++ packages/integrations/index.js
@@ -1,11 +1,15 @@
+import IntegrationsManager from './components/integrationsManager';
 import Disqus from './components/disqus';
 import GoogleAnalytics from './components/googleAnalytics';
 import GoogleTagManager from './components/googleTagManager';
 import Parsely from './components/parsely';
 import Pico from './components/pico';
 import YoastSchema from './components/yoastSchema';
+// Config.
+import config from './irving.config';
 
 export {
+  IntegrationsManager,
   Disqus,
   GoogleAnalytics,
   GoogleTagManager,
@@ -13,3 +17,5 @@ export {
   Pico,
   YoastSchema,
 };
+
+export default config;