+++ packages/styled-components/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v3.1.0...v3.1.1) (2020-08-09)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+# [3.1.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v3.0.1...v3.1.0) (2020-08-09)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
+## [3.0.1](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v3.0.0...v3.0.1) (2020-08-06)
+
+**Note:** Version bump only for package @irvingjs/styled-components
+
+
+
+
+
 # [3.0.0](https://github.com/alleyinteractive/irving/packages/styled-components/compare/v3.0.0-beta.12...v3.0.0) (2020-08-03)
 
 **Note:** Version bump only for package @irvingjs/styled-components